# aspnetcore-server/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Room Reservation API",
    "description" : "A Simple IP Address API",
    "version" : "0.1.9"
  },
  "servers" : [ {
    "url" : "http://localhost:5000/api"
  } ],
  "paths" : {
...
    "/reservations" : {
      "post" : {
        "tags" : [ "reservations" ],
        "summary" : "Creates a new reservation",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MeetingRoomReservation"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeetingRoomReservation"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict"
          }
        }
      }
    },
    "/reservations/{reservationId}" : {
      "get" : {
        "tags" : [ "reservations" ],
        "summary" : "Returns reservation by ID",
        "parameters" : [ {
          "name" : "reservationId",
          "in" : "path",
          "description" : "ID of a reservation",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeetingRoomReservation"
                }
              }
            }
          },
          "404" : {
            "description" : "Reservation not found"
          }
        }
      },
      "put" : {
        "tags" : [ "reservations" ],
        "summary" : "Updates an existing reservation",
        "parameters" : [ {
          "name" : "reservationId",
          "in" : "path",
          "description" : "ID of a reservation",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MeetingRoomReservation"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Updated"
          },
          "400" : {
            "description" : "Bad request"
          },
          "404" : {
            "description" : "Room not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "reservations" ],
        "summary" : "Deletes a reservation",
        "parameters" : [ {
          "name" : "reservationId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "404" : {
            "description" : "Reservation not found"
          }
        }
      }
    },
    "/users" : {
      "get" : {
        "tags" : [ "users" ],
        "summary" : "Returns a list of users",
        "responses" : {
          "200" : {
            "description" : "An array of users",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "404" : {
            "description" : "User not found"
          }
        }
      },
      "post" : {
        "tags" : [ "users" ],
        "summary" : "Creates a new user",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict"
          }
        }
      }
    },
    "/users/{userId}" : {
      "get" : {
        "tags" : [ "users" ],
        "summary" : "Returns user by ID",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "ID of a user",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "404" : {
            "description" : "User not found"
          }
        }
      },
      "put" : {
        "tags" : [ "users" ],
        "summary" : "Updates an existing user",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "ID of a user",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Updated"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "User not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "users" ],
        "summary" : "Deletes a user",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "404" : {
            "description" : "User not found"
          }
        }
      }
    },
    "/users/email/{userEmail}" : {
      "get" : {
        "tags" : [ "users" ],
        "summary" : "Returns user by email",
        "parameters" : [ {
          "name" : "userEmail",
          "in" : "path",
          "description" : "email of a user",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "404" : {
            "description" : "User not found"
          }
        }
      }
    },
    "/users/email/{userEmail}/password/{password}" : {
      "get" : {
        "tags" : [ "users" ],
        "summary" : "Returns user by email and passw",
        "parameters" : [ {
          "name" : "userEmail",
          "in" : "path",
          "description" : "email of a user",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "password",
          "in" : "path",
          "description" : "password of a user",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "404" : {
            "description" : "User not found"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "MeetingRoom" : {
        "type" : "object",
        "properties" : {
          "meetingRoomId" : {
            "type" : "integer",
            "readOnly" : true
          },
          "meetingRoomName" : {
            "type" : "string"
          },
          "meetingRoomEmail" : {
            "type" : "string",
            "format" : "email"
          },
          "meetingRoomLocation" : {
            "$ref" : "#/components/schemas/Location"
          },
          "locationId" : {
            "type" : "integer"
          },
          "meetingRoomImage" : {
            "$ref" : "#/components/schemas/MeetingRoomImage"
          },
          "meetingRoomReservations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MeetingRoomReservation"
            }
          }
        },
        "example" : {
          "meetingRoomLocation" : {
            "meetingRooms" : [ null, null ],
            "city" : {
              "cityName" : "cityName",
              "cityId" : 1
            },
            "locationId" : 6,
            "workingSpots" : [ {
              "workingSpotId" : 5,
              "locationId" : 5,
              "location" : null,
              "workingSpotReservations" : [ {
                "workingSpotId" : 7,
                "workingSpot" : null,
                "workingSpotReservationId" : 2,
                "reservationDate" : "2000-01-23",
                "user" : {
                  "jobId" : 2,
                  "password" : "",
                  "reservations" : [ {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  }, {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  } ],
                  "surname" : "surname",
                  "name" : "name",
                  "job" : {
                    "jobId" : 3,
                    "jobTitle" : "jobTitle"
                  },
                  "userId" : 9,
                  "email" : ""
                },
                "userId" : 1
              }, {
                "workingSpotId" : 7,
                "workingSpot" : null,
                "workingSpotReservationId" : 2,
                "reservationDate" : "2000-01-23",
                "user" : {
                  "jobId" : 2,
                  "password" : "",
                  "reservations" : [ {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  }, {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  } ],
                  "surname" : "surname",
                  "name" : "name",
                  "job" : {
                    "jobId" : 3,
                    "jobTitle" : "jobTitle"
                  },
                  "userId" : 9,
                  "email" : ""
                },
                "userId" : 1
              } ]
            }, {
              "workingSpotId" : 5,
              "locationId" : 5,
              "location" : null,
              "workingSpotReservations" : [ {
                "workingSpotId" : 7,
                "workingSpot" : null,
                "workingSpotReservationId" : 2,
                "reservationDate" : "2000-01-23",
                "user" : {
                  "jobId" : 2,
                  "password" : "",
                  "reservations" : [ {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  }, {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  } ],
                  "surname" : "surname",
                  "name" : "name",
                  "job" : {
                    "jobId" : 3,
                    "jobTitle" : "jobTitle"
                  },
                  "userId" : 9,
                  "email" : ""
                },
                "userId" : 1
              }, {
                "workingSpotId" : 7,
                "workingSpot" : null,
                "workingSpotReservationId" : 2,
                "reservationDate" : "2000-01-23",
                "user" : {
                  "jobId" : 2,
                  "password" : "",
                  "reservations" : [ {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  }, {
                    "meetingRoom" : null,
                    "meetingRoomReservationId" : 4,
                    "meetingRoomId" : 1,
                    "meetingDate" : "2000-01-23",
                    "startTime" : "startTime",
                    "endTime" : "endTime",
                    "meetingTopic" : "meetingTopic",
                    "user" : null,
                    "userId" : 7
                  } ],
                  "surname" : "surname",
                  "name" : "name",
                  "job" : {
                    "jobId" : 3,
                    "jobTitle" : "jobTitle"
                  },
                  "userId" : 9,
                  "email" : ""
                },
                "userId" : 1
              } ]
            } ]
          },
          "meetingRoomId" : 0,
          "meetingRoomName" : "meetingRoomName",
          "locationId" : 1,
          "meetingRoomImage" : {
            "imageWidth" : 7.457744773683766,
            "image" : "",
            "meetingRoom" : null,
            "meetingRoomId" : 1,
            "name" : "name",
            "meetingRoomImageId" : 6
          },
          "meetingRoomReservations" : [ null, null ],
          "meetingRoomEmail" : ""
        }
      },
      "MeetingRoomReservation" : {
        "type" : "object",
        "properties" : {
          "meetingRoomReservationId" : {
            "type" : "integer",
            "readOnly" : true
          },
          "meetingDate" : {
            "type" : "string",
            "format" : "date"
          },
          "startTime" : {
            "type" : "string",
            "format" : "time"
          },
          "endTime" : {
            "type" : "string",
            "format" : "time"
          },
          "meetingTopic" : {
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/User"
          },
          "userId" : {
            "type" : "integer"
          },
          "meetingRoom" : {
            "$ref" : "#/components/schemas/MeetingRoom"
          },
          "meetingRoomId" : {
            "type" : "integer"
          }
        },
        "example" : {
          "meetingRoom" : null,
          "meetingRoomReservationId" : 4,
          "meetingRoomId" : 1,
          "meetingDate" : "2000-01-23",
          "startTime" : "startTime",
          "endTime" : "endTime",
          "meetingTopic" : "meetingTopic",
          "user" : null,
          "userId" : 7
        }
      },
      "User" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "integer",
            "readOnly" : true
          },
          "name" : {
            "type" : "string"
          },
          "surname" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string",
            "format" : "email"
          },
          "password" : {
            "type" : "string",
            "format" : "password"
          },
          "job" : {
            "$ref" : "#/components/schemas/Job"
          },
          "jobId" : {
            "type" : "integer"
          },
          "reservations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MeetingRoomReservation"
            }
          }
        },
        "example" : {
          "jobId" : 2,
          "password" : "",
          "reservations" : [ {
            "meetingRoom" : null,
            "meetingRoomReservationId" : 4,
            "meetingRoomId" : 1,
            "meetingDate" : "2000-01-23",
            "startTime" : "startTime",
            "endTime" : "endTime",
            "meetingTopic" : "meetingTopic",
            "user" : null,
            "userId" : 7
          }, {
            "meetingRoom" : null,
            "meetingRoomReservationId" : 4,
            "meetingRoomId" : 1,
            "meetingDate" : "2000-01-23",
            "startTime" : "startTime",
            "endTime" : "endTime",
            "meetingTopic" : "meetingTopic",
            "user" : null,
            "userId" : 7
          } ],
          "surname" : "surname",
          "name" : "name",
          "job" : {
            "jobId" : 3,
            "jobTitle" : "jobTitle"
          },
          "userId" : 9,
          "email" : ""
        }
      },
      "MeetingRoomImage" : {
        "type" : "object",
        "properties" : {
          "meetingRoomImageId" : {
            "type" : "integer",
            "readOnly" : true
          },
          "name" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string",
            "format" : "byte"
          },
          "imageWidth" : {
            "type" : "number"
          },
          "meetingRoomId" : {
            "type" : "integer"
          },
          "meetingRoom" : {
            "$ref" : "#/components/schemas/MeetingRoom"
          }
        },
        "example" : {
          "imageWidth" : 7.457744773683766,
          "image" : "",
          "meetingRoom" : null,
          "meetingRoomId" : 1,
          "name" : "name",
          "meetingRoomImageId" : 6
        }
      },
...
      "Job" : {
        "type" : "object",
        "properties" : {
          "jobId" : {
            "type" : "integer"
          },
          "jobTitle" : {
            "type" : "string"
          }
        },
        "example" : {
          "jobId" : 3,
          "jobTitle" : "jobTitle"
        }
      },
...
    }
  }
}